switched to u8 instead of u32

+++ src/sha1.d.ts
@@ -1,7 +1,7 @@
 export default function(bytes?: BufferSource): Promise<void>;
 export const wasm: WebAssembly.WebAssemblyInstantiatedSource;
 export const buffer: Uint8Array;
-export const state: Uint32Array;
+export const state: Uint8Array;
 export function init(): void;
 export function hash(length: number): void;
 export function finalize(): void;
